Collapse Karpenter detection into a single FindKarpenterInstallation
Drop the ForeignKarpenter struct and the FindAnyKarpenterInstallation /
FindForeignKarpenterInstallation pair in favour of one
FindKarpenterInstallation that returns the first Karpenter controller it
finds (assuming at most one per cluster, which our install/update guards
enforce). Callers use IsOwn to decide:

- install proceeds when no Karpenter is found, or it is ours.
- update proceeds only when our Karpenter is found.

Update consequently drops its second post-namespace foreign scan, and the
RunOptions skip flag is renamed from SkipForeignKarpenterCheck to
SkipKarpenterCheck. The file is renamed from foreignkarpenter.go to
karpenter.go now that "foreign" is no longer a discriminator.
